test: Import map component to home layout component
