+++ client/src/pages/HomePageElder.tsx
@@ -2,7 +2,6 @@ import { connect } from 'react-redux';
 import { accessTokenState, ProfileState } from '~/others/store';
 import HeaderElder from '~/components/organisms/HeaderElder';
 import HomeElder from '~/components/organisms/HomeLayoutElder';
-import MyMap from '~/components/organisms/MyMap';
 
 interface HomeProps {
   accessToken: accessTokenState;
@@ -14,7 +13,6 @@ const HomePageElder = ({ accessToken, profileData }: HomeProps) => {
     <>
       <HeaderElder accessToken={accessToken} profileData={profileData} />
       <HomeElder />
-      <MyMap />
     </>
   );
 };
